pyproject.toml
[project]
name = "pdfhunter"
version = "0.1.0"
description = "Automatic bibliographic metadata extraction from PDF documents"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "PDFHunter Team"}
]
dependencies = [
    "pypdf>=4.0",
    "pdfplumber>=0.10",
    "pdf2image>=1.16",
    "Pillow>=10.0",
    "opencv-python>=4.8",
    "paddleocr>=2.7",
    "paddlepaddle>=2.5",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "typer>=0.9",
    "openai>=1.0",
    "anthropic>=0.18",
    "rich>=13.0",
    "scholarly>=1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]
ui = [
    "streamlit>=1.30",
]
tesseract = [
    "pytesseract>=0.3",
]

[project.scripts]
pdfhunter = "pdfhunter.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdfhunter"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: marks tests as integration tests (require external dependencies)",
]
